project1/requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
goodreads
requests
